Completed working solution for merge sort.
Fixed the merge sort function.

Started by moving the index increment line after the lines that add the number to the combined array for both the right and left side. This way, the items would be placed staring at the 0th index.

Then I added while loops  to ensure the rest of the left and right arrays get added to the end of the combined array after looking through both arrays. This makes sure that all values are being transfered over.

Lastly, I moved the left index incrementor to the if statement, and the right index incrementor to the else statement. This is because the left index should increase if that current value is the smaller of the two.